requirements.txt
google-generativeai
python-dotenv
openai-whisper @ git+https://github.com/openai/whisper.git
torch
gtts
pathlib
Flask